move data module utility functions from lib.py to data/utils.py
